+++ sdk/ml/azure-ai-ml/azure/ai/ml/_ml_client.py
@@ -73,6 +73,7 @@ from azure.ai.ml.entities import (
     Workspace,
 )
 from azure.ai.ml.entities._assets import WorkspaceAssetReference
+from azure.ai.ml.entities._workspace._ai_workspaces.capability_host import CapabilityHost
 from azure.ai.ml.exceptions import ErrorCategory, ErrorTarget, ValidationException
 from azure.ai.ml.operations import (
     AzureOpenAIDeploymentOperations,
@@ -94,6 +95,7 @@ from azure.ai.ml.operations import (
     ServerlessEndpointOperations,
     WorkspaceConnectionsOperations,
     WorkspaceOperations,
+    CapabilityHostsOperations,
 )
 from azure.ai.ml.operations._code_operations import CodeOperations
 from azure.ai.ml.operations._feature_set_operations import FeatureSetOperations
@@ -525,6 +527,16 @@ class MLClient:
             self._credential,
         )
 
+        self._capability_hosts = CapabilityHostsOperations(
+            self._operation_scope,
+            self._operation_config,
+            self._service_client_10_2024_preview,
+            self._operation_container,
+            self._credential,
+            **kwargs,
+        )
+        self._operation_container.add(AzureMLResourceType.CAPABILITY_HOST, self._capability_hosts)
+
         self._preflight = get_deployments_operation(
             credentials=self._credential,
             subscription_id=self._operation_scope._subscription_id,
@@ -911,6 +923,16 @@ class MLClient:
         :rtype: ~azure.ai.ml.operations.WorkspaceOperations
         """
         return self._workspaces
+
+    @property
+    @experimental
+    def capability_hosts(self) -> CapabilityHostsOperations:
+        """A collection of capability hosts related operations.
+
+        :return: Capability hosts operations
+        :rtype: ~azure.ai.ml.operations.CapabilityHostsOperations
+        """
+        return self._capability_hosts
 
     @property
     def workspace_outbound_rules(self) -> WorkspaceOutboundRuleOperations:
@@ -1347,6 +1369,11 @@ def _begin_create_or_update(entity, operations, **kwargs):
 def _(entity: Workspace, operations, *args, **kwargs):
     module_logger.debug("Creating or updating workspaces")
     return operations[AzureMLResourceType.WORKSPACE].begin_create(entity, **kwargs)
+
+@_create_or_update.register(CapabilityHost)
+def _(entity: CapabilityHost, operations, *args, **kwargs):
+    module_logger.debug("Creating or updating capability host")
+    return operations[AzureMLResourceType.CAPABILITY_HOST].create_or_update(entity, **kwargs)
 
 
 @_begin_create_or_update.register(Registry)