+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### Features Added
   - Add support for additional include in spark component.
+  - Adding enterprise agent feature for AI Hub or AI Project. Adding Create and Get operations.
 
 ### Bugs Fixed